+++ core/tests/ts-integration/src/helpers.ts
@@ -233,3 +233,21 @@ export function bigIntMax(...args: bigint[]) {
 export function isLocalHost(network: string): boolean {
     return network.toLowerCase() == 'localhost';
 }
+
+export function maxL2GasLimitForPriorityTxs(maxGasBodyLimit: bigint): bigint {
+    // Find maximum `gasLimit` that satisfies `txBodyGasLimit <= CONTRACTS_PRIORITY_TX_MAX_GAS_LIMIT`
+    // using binary search.
+    const overhead = getOverheadForTransaction(
+        // We can just pass 0 as `encodingLength` because the overhead for the transaction's slot
+        // will be greater than `overheadForLength` for a typical transacction
+        0n
+    );
+    return maxGasBodyLimit + overhead;
+}
+
+export function getOverheadForTransaction(encodingLength: bigint): bigint {
+    const TX_SLOT_OVERHEAD_GAS = 10_000n;
+    const TX_LENGTH_BYTE_OVERHEAD_GAS = 10n;
+
+    return bigIntMax(TX_SLOT_OVERHEAD_GAS, TX_LENGTH_BYTE_OVERHEAD_GAS * encodingLength);
+}
